Introduced ApplicationSecurity classes and tidied up Application classes (#210)
