+++ forgerock-openbanking-aspsp/forgerock-openbanking-aspsp-rs/forgerock-openbanking-rs-simulator/src/main/java/com/forgerock/openbanking/aspsp/rs/simulator/SimulatorApplication.java
@@ -20,30 +20,29 @@
  */
 package com.forgerock.openbanking.aspsp.rs.simulator;
 
-import com.forgerock.openbanking.common.EnableCookieWebSecurityConfiguration;
-import com.forgerock.openbanking.common.EnableSslClientConfiguration;
+import com.forgerock.openbanking.common.CookieWebSecurityConfiguration;
+import com.forgerock.openbanking.common.EnableSslClient;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.forgerock")
+@EnableMongoRepositories(basePackages = "com.forgerock")
 @EnableSwagger2
 @EnableDiscoveryClient
 @EnableScheduling
-@ComponentScan(basePackages = {"com.forgerock"})
-@EnableMongoRepositories(basePackages = "com.forgerock")
-@EnableCookieWebSecurityConfiguration
-@EnableSslClientConfiguration
+@EnableSslClient
+@Import(CookieWebSecurityConfiguration.class)
 public class SimulatorApplication {
 
     /**
 	 * Set to 'false' to stop any of the scheduled tasks running. Defaults to 'true' if property not present.
 	 */
-	public static final String RUN_SCHEDULED_TASK_PROPERTY = "simulator.runScheduledTasks";
+	public static final String RUN_SCHEDULED_TASK_PROPERTY = "simulator.runScheduledTasks"; // TODO is this used somewhere?
 
 	public static void main(String[] args) {
         SpringApplication.run(SimulatorApplication.class, args);
