--- forgerock-openbanking-common/src/main/java/com/forgerock/openbanking/common/EnableSslClientConfiguration.java
@@ -1,36 +1,0 @@
-/**
- * Copyright 2019 ForgeRock AS.
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.forgerock.openbanking.common;
-
-import org.springframework.context.annotation.Import;
-
-import java.lang.annotation.*;
-
-/**
- * An annotation that can be applied to any Spring Boot application that requires SSL related client configurations, including a {@link org.springframework.web.client.RestTemplate}.
- */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Inherited
-@Import(SslClientConfiguration.class)
-public @interface EnableSslClientConfiguration {
-}
