+++ forgerock-openbanking-directory-services/src/main/java/com/forgerock/openbanking/directory/ForgerockOpenbankingDirectoryApplication.java
@@ -21,79 +21,36 @@
 package com.forgerock.openbanking.directory;
 
 import brave.Tracer;
-import com.forgerock.openbanking.common.EnableSslClientConfiguration;
-import com.forgerock.openbanking.common.OBRIInternalCertificates;
+import com.forgerock.openbanking.common.EnableSslClient;
 import com.forgerock.openbanking.directory.error.ErrorHandler;
 import com.forgerock.openbanking.directory.security.FormValueSanitisationFilter;
 import com.forgerock.openbanking.directory.security.JsonRequestSanitisiationFilter;
-import com.forgerock.openbanking.directory.service.DirectoryUtilsService;
-import com.forgerock.openbanking.jwt.services.CryptoApiClient;
-import com.forgerock.openbanking.ssl.services.keystore.KeyStoreService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 import javax.servlet.Filter;
-import java.security.cert.X509Certificate;
 
-import static com.forgerock.openbanking.common.CookieHttpSecurityConfiguration.configureHttpSecurity;
-
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.forgerock")
+@EnableMongoRepositories(basePackages = "com.forgerock")
 @EnableSwagger2
 @EnableDiscoveryClient
 @EnableScheduling
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
-@ComponentScan(basePackages = {"com.forgerock"})
-@EnableMongoRepositories(basePackages = "com.forgerock")
-@EnableSslClientConfiguration
+@EnableSslClient
+@Import(DirectoryApplicationSecurityConfiguration.class)
 public class ForgerockOpenbankingDirectoryApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(ForgerockOpenbankingDirectoryApplication.class, args);
-    }
-
-    @Configuration
-    static class AuthWebSecurity extends WebSecurityConfigurerAdapter {
-
-        @Value("${matls.forgerock-internal-ca-alias}")
-        private String internalCaAlias;
-        @Value("${matls.forgerock-external-ca-alias}")
-        private String externalCaAlias;
-
-        private final CryptoApiClient cryptoApiClient;
-        private final DirectoryUtilsService directoryUtilsService;
-        private final KeyStoreService keyStoreService;
-
-        @Autowired
-        AuthWebSecurity(CryptoApiClient cryptoApiClient, DirectoryUtilsService directoryUtilsService, KeyStoreService keyStoreService) {
-            this.cryptoApiClient = cryptoApiClient;
-            this.directoryUtilsService = directoryUtilsService;
-            this.keyStoreService = keyStoreService;
-        }
-
-        @Override
-        protected void configure(HttpSecurity http) throws Exception {
-            X509Certificate internalCACertificate = (X509Certificate) keyStoreService.getKeyStore().getCertificate(internalCaAlias);
-            X509Certificate externalCACertificate = (X509Certificate) keyStoreService.getKeyStore().getCertificate(externalCaAlias);
-
-            OBRIInternalCertificates obriInternalCertificates = new OBRIInternalCertificates(internalCACertificate);
-            DirectoryORBIExternalCertificates obriExternalCertificates = new DirectoryORBIExternalCertificates(externalCACertificate, directoryUtilsService);
-
-            configureHttpSecurity(http, obriInternalCertificates, obriExternalCertificates, cryptoApiClient);
-        }
     }
 
     @Bean
