+++ forgerock-openbanking-jwkms/src/main/java/com/forgerock/openbanking/jwkms/ForgerockOpenbankingJwkMSApplication.java
@@ -20,78 +20,32 @@
  */
 package com.forgerock.openbanking.jwkms;
 
-import com.forgerock.openbanking.common.EnableSslClientConfiguration;
-import com.forgerock.openbanking.jwkms.service.application.ApplicationService;
-import com.forgerock.openbanking.jwt.services.CryptoApiClient;
-import com.forgerock.openbanking.ssl.services.keystore.KeyStoreService;
+import com.forgerock.openbanking.common.EnableSslClient;
 import com.mongodb.MongoClient;
-import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.core.LockProvider;
 import net.javacrumbs.shedlock.provider.mongo.MongoLockProvider;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-import java.security.cert.X509Certificate;
-
-import static com.forgerock.openbanking.common.CookieHttpSecurityConfiguration.configureHttpSecurity;
-
-
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.forgerock")
+@EnableMongoRepositories(basePackages = "com.forgerock")
 @EnableSwagger2
 @EnableDiscoveryClient
 @EnableScheduling
 @EnableGlobalMethodSecurity(prePostEnabled = true)
-@Slf4j
-@ComponentScan(basePackages = {"com.forgerock"})
-@EnableMongoRepositories(basePackages = "com.forgerock")
-@EnableSslClientConfiguration
+@EnableSslClient
+@Import(JwkMSApplicationSecurityConfiguration.class)
 public class ForgerockOpenbankingJwkMSApplication  {
 
     public static void main(String[] args) throws Exception {
         new SpringApplication(ForgerockOpenbankingJwkMSApplication.class).run(args);
-    }
-
-    @Configuration
-    static class CookieWebSecurityConfigurerAdapter extends WebSecurityConfigurerAdapter {
-
-        @Value("${matls.forgerock-internal-ca-alias}")
-        private String internalCaAlias;
-        @Value("${matls.forgerock-external-ca-alias}")
-        private String externalCaAlias;
-
-        private final ApplicationService applicationService;
-        private final KeyStoreService keyStoreService;
-        private final CryptoApiClient cryptoApiClient;
-
-        @Autowired
-        CookieWebSecurityConfigurerAdapter(ApplicationService applicationService, KeyStoreService keyStoreService, CryptoApiClient cryptoApiClient) {
-            this.applicationService = applicationService;
-            this.keyStoreService = keyStoreService;
-            this.cryptoApiClient = cryptoApiClient;
-        }
-
-        @Override
-        protected void configure(HttpSecurity http) throws Exception {
-            X509Certificate internalCACertificate = (X509Certificate) keyStoreService.getKeyStore().getCertificate(internalCaAlias);
-            X509Certificate externalCACertificate = (X509Certificate) keyStoreService.getKeyStore().getCertificate(externalCaAlias);
-
-            JwkMsOBRIInternalCertificates obriInternalCertificates = new JwkMsOBRIInternalCertificates(internalCACertificate, applicationService);
-            JwkMsOBRIExternalCertificates obriExternalCertificates = new JwkMsOBRIExternalCertificates(externalCACertificate, applicationService);
-
-            configureHttpSecurity(http, obriInternalCertificates, obriExternalCertificates, cryptoApiClient);
-        }
     }
 
     @Bean
